+++ fili-core/src/main/java/com/yahoo/bard/webservice/data/config/table/BaseTableLoader.java
@@ -319,7 +319,7 @@ public abstract class BaseTableLoader implements TableLoader {
         );
 
         return new ConcretePhysicalTable(
-                definition.getName().asName(),
+                definition.getName(),
                 definition.getGrain(),
                 columns,
                 definition.getLogicalToPhysicalNames()