+++ fili-core/src/main/java/com/yahoo/bard/webservice/table/BasePhysicalTable.java
@@ -2,6 +2,7 @@
 // Licensed under the terms of the Apache license. Please see LICENSE.md file distributed with this work for terms.
 package com.yahoo.bard.webservice.table;
 
+import com.yahoo.bard.webservice.data.config.names.TableName;
 import com.yahoo.bard.webservice.data.dimension.DimensionDictionary;
 import com.yahoo.bard.webservice.data.time.ZonedTimeGrain;
 import com.yahoo.bard.webservice.druid.model.query.Granularity;
@@ -29,7 +30,7 @@ import javax.validation.constraints.NotNull;
 public abstract class BasePhysicalTable implements PhysicalTable {
     private static final Logger LOG = LoggerFactory.getLogger(PhysicalTable.class);
 
-    private final String name;
+    private final TableName name;
     private final PhysicalTableSchema schema;
     private volatile Availability availability;
 
@@ -48,7 +49,7 @@ public abstract class BasePhysicalTable implements PhysicalTable {
      * @param availability  The availability of columns in this table
      */
     public BasePhysicalTable(
-            @NotNull String name,
+            @NotNull TableName name,
             @NotNull ZonedTimeGrain timeGrain,
             @NotNull Iterable<Column> columns,
             @NotNull Map<String, String> logicalToPhysicalColumnNames,
@@ -85,8 +86,13 @@ public abstract class BasePhysicalTable implements PhysicalTable {
     }
 
     @Override
-    public String getName() {
+    public TableName getTableName() {
         return name;
+    }
+
+    @Override
+    public String getName() {
+        return name.asName();
     }
 
     @Override